+++ Snabble/Core/Cart/CheckoutData.swift
@@ -52,7 +52,7 @@ public enum RawPaymentMethod: String, CaseIterable, Decodable {
     case creditCardVisa         // VISA via Telecash/First Data
     case creditCardMastercard   // MASTERCARD via Telecash/First Data
     case creditCardAmericanExpress // AMERICAN EXPRESS via Telecash/First Data
-    case externalBilling        // external billig, e.g. via an employee id
+    case externalBilling        // external billing, e.g. via an employee id
     case gatekeeperTerminal
     case customerCardPOS        // payment via customer card invoice
     case paydirektOneKlick
@@ -60,7 +60,8 @@ public enum RawPaymentMethod: String, CaseIterable, Decodable {
     /// true if this method reqires additional data, like an IBAN or a credit card number
     var dataRequired: Bool {
         switch self {
-        case .deDirectDebit, .creditCardVisa, .creditCardMastercard, .creditCardAmericanExpress, .externalBilling, .customerCardPOS, .paydirektOneKlick:
+        case .deDirectDebit, .creditCardVisa, .creditCardMastercard, .creditCardAmericanExpress,
+             .externalBilling, .customerCardPOS, .paydirektOneKlick:
             return true
         case .qrCodePOS, .qrCodeOffline, .gatekeeperTerminal:
             return false
@@ -260,7 +261,46 @@ public struct CheckoutInfo: Decodable {
     }
 }
 
-/// Checkout Process
+// MARK: - process checks
+public enum CheckState: String, Codable {
+    case unknown
+
+    case pending
+    case successful
+    case failed
+}
+
+extension CheckState: UnknownCaseRepresentable {
+    public static let unknownCase = CheckState.unknown
+}
+
+public enum CheckType: String, Codable {
+    case minAge = "min_age"
+}
+
+public struct CheckoutCheck: Decodable {
+    public let id: String
+    public let links: CheckLinks
+
+    public let state: CheckState
+    public let type: CheckType
+
+    // type-specific properties
+    public let requiredAge: Int? // set for min_age
+
+    public struct CheckLinks: Decodable {
+        public let `self`: Link
+    }
+}
+
+public struct AgeCheckData: Encodable {
+    public let requiredAge: Int
+    public let state: CheckState
+    public let type: CheckType
+    public let dayOfBirth: String // YYYY/MM/DD
+}
+
+// MARK: - Checkout Process
 public struct CheckoutProcess: Decodable {
     public let links: ProcessLinks
     public let supervisorApproval: Bool?
@@ -273,6 +313,7 @@ public struct CheckoutProcess: Decodable {
     public let paymentState: PaymentState
     public let orderID: String?
     public let paymentResult: [String: Any]?
+    public let checks: [CheckoutCheck]
 
     public struct ProcessLinks: Decodable {
         public let `self`: Link
@@ -289,6 +330,7 @@ public struct CheckoutProcess: Decodable {
         case links, supervisorApproval, paymentApproval, aborted
         case checkoutInfo, paymentMethod, modified, paymentInformation
         case paymentState, orderID, paymentResult
+        case checks
     }
 
     public init(from decoder: Decoder) throws {
@@ -305,6 +347,10 @@ public struct CheckoutProcess: Decodable {
         self.paymentState = try container.decode(PaymentState.self, forKey: .paymentState)
         self.orderID = try container.decodeIfPresent(String.self, forKey: .orderID)
         self.paymentResult = try container.decodeIfPresent([String: Any].self, forKey: .paymentResult)
+
+        let rawChecks = try container.decodeIfPresent([FailableDecodable<CheckoutCheck>].self, forKey: .checks)
+        let checks = rawChecks?.compactMap { $0.value } ?? []
+        self.checks = checks.filter { $0.state != .unknown }
     }
 }