+++ Snabble/Core/Metadata/Metadata.swift
@@ -64,6 +64,7 @@ public struct GatewayCertificate: Decodable {
 
 public struct MetadataLinks: Decodable {
     public let clientOrders: Link?
+    public let appUser: Link
     public let appUserOrders: Link
     public let telecashSecret: Link?
     public let telecashPreauth: Link?
@@ -73,6 +74,7 @@ public struct MetadataLinks: Decodable {
 
     fileprivate init() {
         self.clientOrders = nil
+        self.appUser = Link.empty
         self.appUserOrders = Link.empty
         self.createAppUser = Link.empty
         self.`self` = Link.empty
@@ -565,8 +567,12 @@ public extension Metadata {
     static func load(from url: String, completion: @escaping (Metadata?) -> Void ) {
         let project = Project.none
         project.request(.get, url, jwtRequired: false, timeout: 5) { request in
-            guard let request = request, let absoluteString = request.url?.absoluteString else {
+            guard var request = request, let absoluteString = request.url?.absoluteString else {
                 return completion(nil)
+            }
+
+            if SnabbleAPI.debugMode {
+                request.cachePolicy = .reloadIgnoringCacheData
             }
 
             project.perform(request, returnRaw: true) { (result: Result<Metadata, SnabbleError>, raw: [String: Any]?, _) in